Mute Checker Instance Check

+++ sigma/modules/moderation/messages/mute_checker.py
@@ -3,15 +3,16 @@ import discord
 
 async def mute_checker(ev, message):
     if message.guild:
-        if message.author.id not in ev.bot.cfg.dsc.owners:
-            if not message.author.permissions_in(message.channel).administrator:
-                mute_list = ev.db.get_guild_settings(message.guild.id, 'MutedUsers')
-                if mute_list is None:
-                    mute_list = []
-                if message.author.id in mute_list:
-                    try:
-                        await message.delete()
-                    except discord.Forbidden:
-                        pass
-                    except discord.NotFound:
-                        pass
+        if isinstance(message.author, discord.Member):
+            if message.author.id not in ev.bot.cfg.dsc.owners:
+                if not message.author.permissions_in(message.channel).administrator:
+                    mute_list = ev.db.get_guild_settings(message.guild.id, 'MutedUsers')
+                    if mute_list is None:
+                        mute_list = []
+                    if message.author.id in mute_list:
+                        try:
+                            await message.delete()
+                        except discord.Forbidden:
+                            pass
+                        except discord.NotFound:
+                            pass
